ai_service/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
chromadb
pydantic
python-multipart
requests
python-dotenv
dashscope
httpx
sse-starlette
